+++ src/worker/filters/status.rs
@@ -8,7 +8,7 @@ response_filter!(
     needs_body = false,
     |res: &RwalkResponse, range: &Vec<IntRange<u16>>| Ok(range
         .iter()
-        .any(|r| r.contains(res.status))),
+        .any(|r| r.contains(res.status as u16))),
     "status",
     ["code", "s"],
     transform = |raw: String| raw.split(',').map(|s| s.parse()).collect::<Result<_>>()
